Move bot src
