--- commands.js
@@ -1,8 +1,0 @@
-function startMarathon(bot) {
-
-}
-
-
-module.exports = {
-    startMarathon
-}
