+++ src/shared/variables.js
@@ -1,0 +1,7 @@
+const TIME_ZONE = 'Europe/Moscow'
+const DATE_FORMAT = 'YYYY-MM-DD'
+
+module.exports = {
+    TIME_ZONE,
+    DATE_FORMAT,
+}
